Ajustes em Enum de todas as classes. Trait Enum criada. Tratamento de…

+++ app/Services/Tenant/Financial/PaymentTerm/PaymentTermService.php
@@ -20,6 +20,8 @@ class PaymentTermService
 
   public function destroy(int $id): bool
   {
+    // Bloquear exclusão de registros que são necessários para uso do sistema
+    throw_if(($id <= 8), new \Exception(trans('message_lang.this_is_system_control_cannot_delete')));
     return $this->repository->destroy($id);
   }
